Add admin layout and pages for dashboard, students, and settings
- Created a new AdminLayout component for consistent sidebar navigation and layout structure.
- Implemented the AdminDashboard page with stats and recent activities.
- Added a StudentsPage for managing student information with a table view.
- Introduced a SettingsPage with a placeholder for future settings functionality.
